nop_bai_tap_ngay19_thang6

+++ module2/src/ss12_java_collection_framework/bai_tap/luyen_tap_arraylist_linkedlist/Product.java
@@ -3,12 +3,12 @@ package ss12_java_collection_framework.bai_tap.luyen_tap_arraylist_linkedlist;
 public class Product {
     public int id;
     public String name;
-    public int price;
+    public double price;
 
     public Product() {
     }
 
-    public Product(int id, String name, int price) {
+    public Product(int id, String name, double price) {
         this.id = id;
         this.name = name;
         this.price = price;
@@ -30,7 +30,7 @@ public class Product {
         this.name = name;
     }
 
-    public int getPrice() {
+    public double getPrice() {
         return price;
     }
 
@@ -46,4 +46,5 @@ public class Product {
                 ", price=" + price +
                 '}';
     }
+
 }
